- Remove tau.cpp, tau.h, put tau step calculations in tau_leaper.cpp for faster performance (subprocess is slowed down when calling methods from multiple files), increases speed on smaller models (michaelis menten, vilar oscillator) by .3-.5 seconds.
